+++ libsolidity/ast/Types.cpp
@@ -2537,6 +2537,7 @@ Type const& UserDefinedValueType::underlyingType() const
 {
 	Type const* type = m_definition.underlyingType()->annotation().type;
 	solAssert(type, "");
+	solAssert(type->category() != Category::UserDefinedValueType, "");
 	return *type;
 }
 
@@ -3071,10 +3072,7 @@ u256 FunctionType::storageSize() const
 
 bool FunctionType::leftAligned() const
 {
-	if (m_kind == Kind::External)
-		return true;
-	else
-		solAssert(false, "Alignment property of non-exportable function type requested.");
+	return m_kind == Kind::External;
 }
 
 unsigned FunctionType::storageBytes() const